refactor: create utils for shrinkwrapping conv
